% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addDiffDays}
\alias{addDiffDays}
\title{Add date difference to data frame}
\usage{
addDiffDays(.data, datesA, datesB, .name = NULL)
}
\arguments{
\item{.data}{Anything accepted by dplyr.}

\item{datesA}{Date column.}

\item{datesB}{Date column.}

\item{.name}{Name of the column.}
}
\description{
Add date difference to data frame
}
