% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{addExclusion}
\alias{addExclusion}
\alias{addExclusion,dataRegister-method}
\title{Add exclusion}
\usage{
addExclusion(.data, ..., description = NULL)

\S4method{addExclusion}{dataRegister}(.data, ..., description = NULL)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Column-value pairs.}

\item{description}{A description of the exclusion}
}
\description{
Add exclusion
}
