% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{addPredecessor}
\alias{addPredecessor}
\title{Add the preceding value of a sequence as a column in data}
\usage{
addPredecessor(.data, colStates, colName = NULL)
}
\arguments{
\item{.data}{Any tabular data accepted by dplyr}

\item{colStates}{The column indicating current state.}

\item{colName}{Name of column for the origin state. Defaults to "from\code{colStates}".}
}
\description{
Add the preceding value of a sequence as a column in data
}
