% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjacent}
\alias{adjacent}
\title{Get adjacent numbers}
\usage{
adjacent(.numbers, .distance = 1)
}
\arguments{
\item{.numbers}{A vector of numbers}

\item{.distance}{Determinant of adjacent numbers.}
}
\description{
Get adjacent numbers
}
