% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{applyExclusions}
\alias{applyExclusions}
\alias{applyExclusions,dataRegister-method}
\title{Remove observations from data marked to be excluded}
\usage{
applyExclusions(.data, verbose = T)

\S4method{applyExclusions}{dataRegister}(.data, verbose = T)
}
\arguments{
\item{.data}{Anything accepted by dplyr.}

\item{verbose}{Informative output.}
}
\description{
Remove observations from data marked to be excluded
}
\seealso{
\code{\link[=addExclusion]{addExclusion()}}

\code{\link[=getExclusions]{getExclusions()}}
}
