% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{closest}
\alias{closest}
\title{Get the number(s) in a vector that are closest to another number.}
\usage{
closest(x, y = 0, na.rm = T)
}
\arguments{
\item{x}{A vector of numbers.}

\item{y}{The comparison number.}

\item{na.rm}{Whether to ignore missing.}
}
\value{
The number(s) in x closest to y.
}
\description{
Get the number(s) in a vector that are closest to another number.
}
