% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseUnique}
\alias{collapseUnique}
\title{Collapse unique elements in a vector}
\usage{
collapseUnique(x, sep = ",", sorted = T)
}
\arguments{
\item{x}{A vector.}

\item{sep}{A character separating elements.}

\item{sorted}{Whether to sort elements.}
}
\description{
Collapse unique elements in a vector
}
