% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configRQCR}
\alias{configRQCR}
\alias{configRQCR,character,character,ANY-method}
\alias{configRQCR,character,character,missing-method}
\title{Configure RQCR}
\usage{
configRQCR(what, setting, value)

\S4method{configRQCR}{character,character,ANY}(what, setting, value)

\S4method{configRQCR}{character,character,missing}(what, setting)
}
\arguments{
\item{what}{What to configure.}

\item{setting}{Which setting.}

\item{value}{Setting value.}
}
\description{
Configure RQCR
}
