% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{configure}
\alias{configure}
\alias{configure,dataRegister,character-method}
\title{Configure a \code{dataRegister} object}
\usage{
configure(.data, file)

\S4method{configure}{dataRegister,character}(.data, file)
}
\arguments{
\item{.data}{A \code{dataRegister} object.}

\item{file}{A yaml file with configuration.}
}
\description{
Configure a \code{dataRegister} object
}
