% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-conversion.R
\name{convertBigCSV}
\alias{convertBigCSV}
\title{Memory efficient way to convert CSV/text files}
\usage{
convertBigCSV(fileInput, fileOutput, nlines = 1e+06, verbose = F, ...)
}
\arguments{
\item{fileInput}{A csv/text file.}

\item{fileOutput}{Output file where type is determined by exension.}

\item{nlines}{Number of lines to read in each chunk of fileInput.}

\item{verbose}{Verbose output?}

\item{...}{Arguments passed to data.table::fread.}
}
\description{
Memory efficient way to convert CSV/text files
}
\details{
File formats are not limited to .csv files, but anything accepted by data.table::fread.
}
