% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-conversion.R
\name{convertDBTableToParquet}
\alias{convertDBTableToParquet}
\title{Convert a database table into a parquet file.}
\usage{
convertDBTableToParquet(dbCon, table, pathOutput, partition = F, overwrite = T)
}
\arguments{
\item{dbCon}{A database connection (needs to support conversion to parquet).}

\item{table}{Database table name.}

\item{pathOutput}{Parquet output path.}

\item{partition}{Whether to partition data overa column.}

\item{overwrite}{Whether to overwrite an existing parquet file.}
}
\description{
Convert a database table into a parquet file.
}
