% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countUnique}
\alias{countUnique}
\title{Count the unique elements in data}
\usage{
countUnique(dta, ...)
}
\arguments{
\item{dta}{A data.frame or similar.}

\item{...}{Columns in \code{dta}.}
}
\description{
Count the unique elements in data
}
