% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{dataRegister}
\alias{dataRegister}
\title{Constructor for register class}
\usage{
dataRegister(x, register)
}
\arguments{
\item{x}{Tabular data.}

\item{register}{The type of register.}
}
\description{
Constructor for register class
}
