% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{dbDuckConnect}
\alias{dbDuckConnect}
\title{Connect to a DuckDB database}
\usage{
dbDuckConnect(dbFile, read_only = T, ...)
}
\arguments{
\item{dbFile}{The database file.}

\item{read_only}{Connect without writing permission?}

\item{...}{Further arguments passed to \code{duckdb::duckdb()}.}
}
\description{
Connect to a DuckDB database
}
\details{
The main purpose of this function is to provide a default with read_only=T
with a database stored in a file. For in-memory databases, you might aswell
just use `DBI::dbConnect(duckdb::duckdb()).
}
