% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{dbDuckConnectMany}
\alias{dbDuckConnectMany}
\title{Connect multiple DuckDB databases}
\usage{
dbDuckConnectMany(..., read_only = T)
}
\arguments{
\item{...}{Named arguments for other databases, \code{database-name = "database-file"}}

\item{read_only}{Whether connections should be in read only mode.}
}
\description{
Connect multiple DuckDB databases
}
\details{
This function sets up a database from multiple DuckDB database files by creating
a "main" in-memory database. Each argument to \code{...} must be named to give
each database a name.
}
