% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbDuckDisconnect}
\alias{dbDuckDisconnect}
\title{Disconnect from a DuckDB.}
\usage{
dbDuckDisconnect(con, checkpoint = T)
}
\arguments{
\item{con}{A database connection.}

\item{checkpoint}{Whether to issue a CHECKPOINT prior to exiting.}
}
\description{
Disconnect from a DuckDB.
}
