% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diffYears}
\alias{diffYears}
\title{Calculate difference between dates in years.}
\usage{
diffYears(datesA, datesB)
}
\arguments{
\item{datesA}{A date/vector of dates.}

\item{datesB}{A date/vector of dates.}
}
\value{
\code{dateA}-\code{dateB} in years.
}
\description{
Calculate difference between dates in years.
}
