% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distanceBetween}
\alias{distanceBetween}
\title{Calculate temporal distance between a set of states}
\usage{
distanceBetween(states, times, from, to)
}
\arguments{
\item{states}{States corresponing to each time in \code{times}.}

\item{times}{Observation times for \code{states}}

\item{from}{The origin state.}

\item{to}{The destination state(s).}
}
\value{
A sequence of time differences with NA if the origin state is not equal
to \code{from} or the destination is not present in \code{to}.
}
\description{
Calculate temporal distance between a set of states
}
\details{
This function calculates the difference between a set of states in time. The
origin state is only one, but the destinations can be many.
}
