% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{drugRegister}
\alias{drugRegister}
\alias{drugRegister,character,character-method}
\alias{drugRegister,character,ANY-method}
\alias{drugRegister,data.frame,character-method}
\title{Create object for the drug register}
\usage{
drugRegister(x, columns)

\S4method{drugRegister}{character,character}(x, columns)

\S4method{drugRegister}{character,ANY}(x)

\S4method{drugRegister}{data.frame,character}(x, columns)
}
\arguments{
\item{x}{Register input: a file or tabular data.}

\item{columns}{Register columns.}
}
\description{
Create object for the drug register
}
