% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elements}
\alias{elements}
\alias{elements,ANY-method}
\alias{elements,data.frame-method}
\title{Get unique elements in data.}
\usage{
elements(.data, ...)

\S4method{elements}{ANY}(.data)

\S4method{elements}{data.frame}(.data, ..., lowercase = F)
}
\arguments{
\item{.data}{Any data accepted by dplyr.}

\item{...}{Columns in .data.}

\item{lowercase}{Lowercase all characters.}
}
\description{
Get unique elements in data.
}
