% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandRange}
\alias{expandRange}
\title{Expand a numerical range embedded in strings}
\usage{
expandRange(x, sep = "-", leadZeroes = T)
}
\arguments{
\item{x}{Strings to expand.}

\item{sep}{A separator for ranges}

\item{leadZeroes}{Whether to align numbers in strings to equal length, eg 1 becomes 01 if range max is within 10-99, eg A9-A10}
}
\description{
Expand a numerical range embedded in strings
}
\details{
\code{expandRange} will take strings in the format of "A#-A#" and repeat these
into their implied range.
}
\examples{
vec <- c("A", "A1-A3")
expandRange(vec)
# Multiple delimiters
expandRange(c('A9/A10', 'B11-B12'), sep=c('/', '-'))

}
