% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{frequencyATC}
\alias{frequencyATC}
\alias{frequencyATC,dataRegister-method}
\title{Calculate frequencies of ATC codes}
\usage{
frequencyATC(x, ...)

\S4method{frequencyATC}{dataRegister}(x, ...)
}
\arguments{
\item{x}{Input data.}

\item{...}{Grouping variables.}
}
\description{
Calculate frequencies of ATC codes
}
