% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frequencyCountDiscrete}
\alias{frequencyCountDiscrete}
\title{Count frequencies of discrete values}
\usage{
frequencyCountDiscrete(.data, ...)
}
\arguments{
\item{.data}{Any data accepted by dplyr.}

\item{...}{Column(s) to form discrete values in \code{dta}.}
}
\value{
A tibble with columns in \code{...} and columns for counts (\code{N}) and percent.
}
\description{
Count frequencies of discrete values
}
\details{
Counting is performed by counting rows in data which forms combinations
of unique values by the specified columns in \code{...}. This is what \code{dplyr::count}
does, but adds a percentage column.
}
