% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{getColumn}
\alias{getColumn}
\title{Get column name using generic name}
\usage{
getColumn(name, type)
}
\arguments{
\item{name}{The generic name.}

\item{type}{Currently "English" or "Norwegian"}
}
\description{
Get column name using generic name
}
