% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{getExclusions}
\alias{getExclusions}
\alias{getExclusions,dataRegister-method}
\title{Get exclusions}
\usage{
getExclusions(.data)

\S4method{getExclusions}{dataRegister}(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\description{
Get exclusions
}
