% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPrefixMatches}
\alias{getPrefixMatches}
\alias{getPrefixMatches,character,character-method}
\title{Get strings that matches a prefix}
\usage{
getPrefixMatches(prefix, str)

\S4method{getPrefixMatches}{character,character}(prefix, str)
}
\arguments{
\item{prefix}{One or several prefix.}

\item{str}{One or several strings.}
}
\description{
Get strings that matches a prefix
}
