% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glueCollapse}
\alias{glueCollapse}
\title{String interpolation with pre-collapsing}
\usage{
glueCollapse(.x, ..., sep = "")
}
\arguments{
\item{.x}{See documentation for \code{glue::glue}.}

\item{...}{Arguments to \code{glue::glue}.}

\item{sep}{Separator of values.}
}
\description{
String interpolation with pre-collapsing
}
\details{
Note that when using this function with variables defined in that function
scope you need to set \code{.envir=enviornment()}.
}
