% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{identifiers}
\alias{identifiers}
\alias{getColumn}
\alias{getColumn,dataRegister,character-method}
\alias{setColumn}
\alias{setColumn,dataRegister,character,character-method}
\title{Get and set identifiers for columns}
\usage{
getColumn(.data, name)

\S4method{getColumn}{dataRegister,character}(.data, name)

setColumn(.data, name, column)

\S4method{setColumn}{dataRegister,character,character}(.data, name, column)
}
\arguments{
\item{.data}{A dataRegister object.}

\item{name}{The identifier name.}

\item{column}{The column name.}
}
\description{
Get and set identifiers for columns

Set identifier column name
}
