% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indexAlong}
\alias{indexAlong}
\title{Index observations over grouped data}
\usage{
indexAlong(dta, ...)
}
\arguments{
\item{dta}{A data frame or similar.}

\item{...}{Columns in \code{dta} defining the groups.}
}
\description{
Index observations over grouped data
}
\details{
Without passing any groups provided, this function just returns an index
(row number) of the data.
}
\seealso{
\code{\link[=indexAlongUnique]{indexAlongUnique()}}
}
