% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indexAlongDistance}
\alias{indexAlongDistance}
\title{Index a sequence of numbers}
\usage{
indexAlongDistance(.numbers, .distance = 1)
}
\arguments{
\item{.numbers}{A vector of numbers.}

\item{.distance}{The distance between two numbers to consider them being adjacent.}
}
\description{
Index a sequence of numbers
}
\details{
\code{indexAlongDistance} index a sequence of numbers setting the same index to
adjacent numbers. Adjacency is determined by \code{.distance}.
}
