% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{indexObservations}
\alias{indexObservations}
\alias{indexObservations,dataRegister-method}
\title{Index observations}
\usage{
indexObservations(.data, ...)

\S4method{indexObservations}{dataRegister}(.data, ..., .nameIndex = "i", .nameMax = NA)
}
\arguments{
\item{.data}{Any data accepted by dplyr.}

\item{...}{Variables to group data by.}

\item{.nameIndex}{Column name.}

\item{.nameMax}{Column name for maximum.}
}
\description{
Index observations
}
