% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{indicateEventsWithin}
\alias{indicateEventsWithin}
\alias{indicateEventsWithin,data.frame,data.frame-method}
\title{Merge two datasets and indicate falling within a date range}
\usage{
indicateEventsWithin(
  .data,
  events,
  dateStart,
  dateEnd,
  dateEvent,
  .by = NULL,
  .slide = 0,
  .ids = NULL,
  .name = "event"
)

\S4method{indicateEventsWithin}{data.frame,data.frame}(
  .data,
  events,
  dateStart,
  dateEnd,
  dateEvent,
  .by = NULL,
  .slide = 0,
  .ids = NULL,
  .name = "event"
)
}
\arguments{
\item{.data}{Anything accepted by dplyr.}

\item{events}{Data with events.}

\item{dateStart}{Period start date column in \code{.data}.}

\item{dateEnd}{Period end date column in \code{.data}.}

\item{dateEvent}{Event date column in \code{events}.}

\item{.by}{Grouping columns.}

\item{.slide}{Integer to shift \code{dateStart} and \code{dateEnd} with.}

\item{.ids}{Ids for joining, adheres to \code{dplyr::join_by}.}

\item{.name}{Column name for event indicator.}
}
\description{
Merge two datasets and indicate falling within a date range
}
