% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{indicateNovel}
\alias{indicateNovel}
\alias{indicateNovel,factor-method}
\alias{indicateNovel,character-method}
\alias{indicateNovel,data.frame-method}
\title{Indicate novel elements in ordered data}
\usage{
indicateNovel(.data, ...)

\S4method{indicateNovel}{factor}(.data)

\S4method{indicateNovel}{character}(.data)

\S4method{indicateNovel}{data.frame}(.data, times, values, ..., .name = "novel")
}
\arguments{
\item{.data}{Any data object accepted by dplyr.}

\item{...}{Grouping factors.}

\item{times}{Column referring to the temporal ordering.}

\item{values}{Column referring to values to mark as new.}

\item{.name}{Name of the column indicating novelty.}
}
\description{
Indicate novel elements in ordered data
}
