% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intersectEquals}
\alias{intersectEquals}
\title{Test if a vector is equal to the intersection of other vectors.}
\usage{
intersectEquals(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Vectors to intersect.}
}
\description{
Test if a vector is equal to the intersection of other vectors.
}
