% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isDate}
\alias{isDate}
\title{Test if x is a date}
\usage{
isDate(x, format)
}
\arguments{
\item{x}{Anything}

\item{format}{If any other format than specified by as.Date}
}
\description{
Test if x is a date
}
