% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{joinPrefix}
\alias{joinPrefix}
\alias{joinLeftPrefix}
\alias{joinLeftPrefix,data.frame,data.frame,dplyr_join_by-method}
\title{Join data based on column as prefixes}
\usage{
joinLeftPrefix(x, y, by, ...)

\S4method{joinLeftPrefix}{data.frame,data.frame,dplyr_join_by}(x, y, by)
}
\arguments{
\item{x}{Any data object accepted by fuzzyjoin.}

\item{y}{Data with prefix column(s),}

\item{by}{Specification of joining, equavalent to \code{dplyr::join_by}. Note: Only equality joins (\code{==}) accepted.}

\item{...}{Other arguments}
}
\description{
This function joins data based on prefixes in a column. If \code{x} is data with
a character column, and \code{y} data with a column of prefixes, \code{joinLeftPrefix}
will perform a join pairs of columns where prefixes in \code{y} matches prefixes in \code{x}.
}
