% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelStrings}
\alias{labelStrings}
\alias{labelStrings,character,character,character,ANY-method}
\alias{labelStrings,character,character,character,missing-method}
\alias{labelStrings,character,list,ANY,ANY-method}
\title{Match a vector against another character set and label matches.}
\usage{
labelStrings(strings, sets, labels, exclude)

\S4method{labelStrings}{character,character,character,ANY}(strings, sets, labels, exclude)

\S4method{labelStrings}{character,character,character,missing}(strings, sets, labels)

\S4method{labelStrings}{character,list,ANY,ANY}(strings, sets)
}
\arguments{
\item{strings}{A vector of strings.}

\item{sets}{A vector of strings to match against.}

\item{labels}{A vector of labels to assign to matches.}

\item{exclude}{A vector of strings to exclude. Either one for all sets or one per set.}
}
\description{
Match a vector against another character set and label matches.
}
