% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{logging}
\alias{logging}
\alias{logExclusion}
\alias{logExclusion,dataRegister,character,numeric-method}
\alias{logCheckpoint}
\alias{logCheckpoint,dataRegister,character,numeric-method}
\alias{logReset}
\alias{logReset,dataRegister-method}
\alias{getLog}
\alias{getLog,dataRegister-method}
\title{Logging}
\usage{
logExclusion(x, what, statistic, description)

\S4method{logExclusion}{dataRegister,character,numeric}(x, what, statistic, description)

logCheckpoint(x, what, statistic, description)

\S4method{logCheckpoint}{dataRegister,character,numeric}(x, what, statistic, description)

logReset(x)

\S4method{logReset}{dataRegister}(x)

getLog(x)

\S4method{getLog}{dataRegister}(x)
}
\arguments{
\item{x}{A \code{dataRegister} object.}

\item{what}{What is excluded?}

\item{statistic}{Number of exclusions.}

\item{description}{A description of the exclusion.}
}
\description{
Functions to log operations on data.
}
