% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maxNegative}
\alias{maxNegative}
\title{Maximum of negative numbers in a vector.}
\usage{
maxNegative(x, na.rm = T)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Whether to ignore missing.}
}
\description{
Maximum of negative numbers in a vector.
}
