% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeDataByRow}
\alias{mergeDataByRow}
\title{Merge data by row like rbind, but with options to separate data.}
\usage{
mergeDataByRow(..., insertColnames = T, separator = NA)
}
\arguments{
\item{...}{Tabular objects like data frames.}

\item{insertColnames}{Insert the column names of each table as the first row in the data.}

\item{separator}{Separate each table with a row of this value (use NA for whitespace).}
}
\value{
A data.frame.
}
\description{
Merge data by row like rbind, but with options to separate data.
}
\details{
Like the motivation for \code{splitDataByRow}, the motivation for this function
comes from spreadsheets where several tables may be present in a sheet.
}
\seealso{
\code{\link[=splitDataByRow]{splitDataByRow()}}
}
