% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{mergeEventsWithin}
\alias{mergeEventsWithin}
\alias{mergeEventsWithin,data.frame,data.frame-method}
\title{Merge two datasets and filter using dates}
\usage{
mergeEventsWithin(
  .data,
  events,
  dateStart,
  dateEnd,
  dateEvent,
  ...,
  .slide = 0,
  ids = NULL
)

\S4method{mergeEventsWithin}{data.frame,data.frame}(
  .data,
  events,
  dateStart,
  dateEnd,
  dateEvent,
  ...,
  .slide = 0,
  ids = NULL
)
}
\arguments{
\item{.data}{Anything accepted by dplyr.}

\item{events}{Data with events.}

\item{dateStart}{Period start date column in \code{.data}.}

\item{dateEnd}{Period end date column in \code{.data}.}

\item{dateEvent}{Event date column in \code{events}.}

\item{...}{Grouping columns.}

\item{.slide}{Integer to shift \code{dateStart} and \code{dateEnd} with.}

\item{.ids}{Ids for joining, adheres to \code{dplyr::join_by}.}
}
\description{
Merge two datasets and filter using dates
}
