% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{mergePeriods}
\alias{mergePeriods}
\title{Merge time periods}
\usage{
mergePeriods(.date, .days)
}
\arguments{
\item{.date}{Period starting date (YYYY-MM-DD)}

\item{.days}{Period length in days}
}
\description{
Merge time periods
}
