% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minPositive}
\alias{minPositive}
\title{Minimum of positive numbers in a vector.}
\usage{
minPositive(x, na.rm = T)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Whether to ignore missing.}
}
\description{
Minimum of positive numbers in a vector.
}
