% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{positionsDistant}
\alias{positionsDistant}
\title{Get positions of distant numbers}
\usage{
positionsDistant(numbers, .distance = 1)
}
\arguments{
\item{numbers}{A vector of numbers}

\item{.distance}{Determinant of distant numbers.}
}
\description{
Get positions of distant numbers
}
