% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdr.R
\name{processDrugRegister}
\alias{processDrugRegister}
\title{Perform QC on the drug register}
\usage{
processDrugRegister(
  fileInput,
  fileOutput,
  formatInput = "csv",
  formatOutput = "csv"
)
}
\arguments{
\item{fileInput}{The drug register.}

\item{fileOutput}{The output file.}

\item{formatInput}{File format of input.}

\item{formatOutput}{File format of output.}
}
\description{
Perform QC on the drug register
}
