% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recodeSex}
\alias{recodeSex}
\title{Recode individual sex to M(ale)/F(emale).}
\usage{
recodeSex(x, default = NA)
}
\arguments{
\item{x}{A vector of inputs.}

\item{default}{Default value of x if no match is made.}
}
\description{
Recode individual sex to M(ale)/F(emale).
}
