% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recodeToNA}
\alias{recodeToNA}
\title{Recode values to missing.}
\usage{
recodeToNA(x, cases)
}
\arguments{
\item{x}{A vector of inputs.}

\item{cases}{Cases in x to convert to NA.}
}
\description{
Recode values to missing.
}
