% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{renameColumns}
\alias{renameColumns}
\alias{renameColumns,dataRegister,character,character-method}
\alias{renameColumns,data.frame,character,character-method}
\title{Rename columns in a \code{dataRegister} object}
\usage{
renameColumns(.data, colsNew, colsRename = NULL, verbose = F)

\S4method{renameColumns}{dataRegister,character,character}(.data, colsNew, colsRename = NULL, verbose = F)

\S4method{renameColumns}{data.frame,character,character}(.data, colsNew, colsRename = NULL, verbose = F)
}
\arguments{
\item{.data}{A \code{dataRegister} object.}

\item{colsNew}{Vector with new column names.}

\item{colsRename}{Vector with column names to change.}

\item{verbose}{Whether to output information.}
}
\description{
Rename columns in a \code{dataRegister} object
}
