% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{renameColumnsMap}
\alias{renameColumnsMap}
\title{Rename columns in data}
\usage{
renameColumnsMap(dta, colsName, colsRename = NULL, verbose = T)
}
\arguments{
\item{dta}{A data.frame.}

\item{colsName}{New column names.}

\item{colsRename}{Columns to rename.}

\item{verbose}{Control information output.}
}
\description{
Rename columns in data
}
