% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{rqcr-columns}
\alias{rqcr-columns}
\alias{getColumn}
\alias{getColumn,dataRegister,character-method}
\alias{setColumn}
\alias{setColumn,dataRegister,character,character-method}
\title{Get/set columns for dataRegister methods}
\usage{
getColumn(.data, name)

\S4method{getColumn}{dataRegister,character}(.data, name)

setColumn(.data, name, column)

\S4method{setColumn}{dataRegister,character,character}(.data, name, column)
}
\arguments{
\item{.data}{A dataRegister object.}

\item{name}{Name of the column identifier.}

\item{column}{The column in data.}
}
\description{
Set and get column identifiers used by \code{dataRegister} class methods. These identifiers
are internal names that refer to columns in the actual data.
}
