% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{save.dataRegister}
\alias{save.dataRegister}
\title{Save data to file}
\usage{
save.dataRegister(x, file, format, ...)
}
\arguments{
\item{x}{A dataRegister class.}

\item{file}{The filename}

\item{format}{The file format.}

\item{...}{Further parameters passed to \code{rio::export}}
}
\description{
Save data to file
}
