% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitVector}
\alias{splitVector}
\title{Split a vector at positions}
\usage{
splitVector(.x, .pos)
}
\arguments{
\item{.x}{A vector.}

\item{.pos}{Positions to split \code{.x} by.}
}
\description{
Split a vector at positions
}
