% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{standardize}
\alias{standardize}
\title{Standardize a variable to mean zero and unit variance}
\usage{
standardize(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Whether to remove missing values.}
}
\description{
Standardize a variable to mean zero and unit variance
}
