% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{standardizeColumns}
\alias{standardizeColumns}
\alias{standardizeColumns,data.frame-method}
\title{Standardize columns in data}
\usage{
standardizeColumns(.data, ..., .names = NULL)

\S4method{standardizeColumns}{data.frame}(.data, ..., .names = NULL)
}
\arguments{
\item{.data}{Anything accepted by dplyr.}

\item{...}{Columns to standardize.}

\item{.names}{Use to name the standardize column see \code{dplyr::across}, E.g \code{"prefix{.col}"}.}
}
\description{
Standardize columns in data
}
