% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txtNPercent}
\alias{txtNPercent}
\title{Create a string with n and percent}
\usage{
txtNPercent(n, N, ..., strmask = "{n} ({percent})", maxPercent = NULL)
}
\arguments{
\item{n}{A number}

\item{N}{Another number}

\item{...}{Arguments to \code{format}}

\item{strmask}{The text to put numbers in. Must contain "\{n\}" and "\{percent\}".}

\item{maxPercent}{If the calculated percent exceeds this limit, a warning is thrown.}
}
\description{
Create a string with n and percent
}
