% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{updateCases}
\alias{updateCases}
\title{Update values conditionally in tabular data}
\usage{
updateCases(.data, col, ..., .warnIfMissing = FALSE)
}
\arguments{
\item{.data}{Anything accepted by dplyr (can be piped).}

\item{col}{The column to update}

\item{...}{Conditions for updates in the form \code{Column == "value" ~ Replacement}.}

\item{.warnIfMissing}{Throws a warning if the update condition is not identified in .data}
}
\description{
Update values conditionally in tabular data
}
\details{
This function uses dplyr::case_when to perform conditional updates in data, but
by default preserving existing values for rows that do not match.
}
\seealso{
\code{\link[dplyr:case_when]{dplyr::case_when()}}
}
