% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whichTransitionsInterval}
\alias{whichTransitionsInterval}
\title{Find adjacent rows where a state goes from one to the next within a time frame}
\usage{
whichTransitionsInterval(states, times, from, to, lower = -Inf, upper = Inf)
}
\arguments{
\item{states}{States corresponing to each time in \code{times}.}

\item{times}{Observation times for \code{states}}

\item{from}{The origin state.}

\item{to}{The destination state(s).}

\item{lower}{Lower end of interval.}

\item{upper}{Upper end of interval.}
}
\description{
Find adjacent rows where a state goes from one to the next within a time frame
}
