% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{addExclusion}
\alias{addExclusion}
\alias{addExclusion,dataRegister-method}
\title{Add exclusion}
\usage{
addExclusion(.data, column, value, description = NULL)

\S4method{addExclusion}{dataRegister}(.data, column, value, description = NULL)
}
\arguments{
\item{.data}{A data frame.}

\item{column}{Column with values to exclude.}

\item{value}{Values in \code{column} to exclude.}

\item{description}{A description of the exclusion}
}
\description{
Add exclusion
}
