% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateAge}
\alias{calculateAge}
\title{Calculate age based on YYYY-MM-DD.}
\usage{
calculateAge(dateBirth, when = today())
}
\arguments{
\item{dateBirth}{Birth date.}

\item{when}{Date when age is to be calculated.}
}
\description{
Calculate age based on YYYY-MM-DD.
}
