% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateAge}
\alias{calculateAge}
\title{Calculate age.}
\usage{
calculateAge(dateBirth, when = today())
}
\arguments{
\item{Birth}{date.}

\item{Date}{when age is to be calculated.}
}
\description{
Calculate age.
}
