% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdr.R
\name{compareDrugFrequencies}
\alias{compareDrugFrequencies}
\title{Compare statistics between the Norwegian drug register and public data.}
\usage{
compareDrugFrequencies(df.local, df.public)
}
\arguments{
\item{df.local}{Drug register data at an individual level}

\item{df.public}{Raw data downloaded from statistikk.fhi.no.}
}
\value{
A dataframe with ATC code, age group, year, sex, and columns Public,
Local, and their difference (Public - Local).
}
\description{
Compare statistics between the Norwegian drug register and public data.
}
\details{
Additional details...
Calculate and compare frequencies by ATC codes in the Norweigan drug register
with aggregated data available through the Norwegian Institute of Public Health's
public API available at statistikk.fhi.no. This data can be downloaded through
R using the package getStatisticsFHI on which this function depends. See
https://github.com/deepchocolate/get-statistics-fhi for details.
}
