% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-conversion.R
\name{convertToCSV}
\alias{convertToCSV}
\title{Convert a file to CSV.}
\usage{
convertToCSV(
  fileInput,
  fileOutput,
  formatInput,
  dropColumns = c(),
  lowerColumns = F
)
}
\arguments{
\item{fileInput}{Path to an input file.}

\item{fileOutput}{Path to the output CSV.}

\item{formatInput}{Specify input format if not provided by file extension.}

\item{dropColumns}{A vector of columns to remove.}

\item{lowerColumns}{Wheter to lowercase column names.}
}
\description{
Convert a file to CSV.
}
\details{
Any file format accepted by rio::convert is accepted as input. If directories in the
output file path do not exist, they are created.
}
