% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{countUnique}
\alias{countUnique}
\alias{countUnique,data.frame-method}
\alias{countUnique,ANY-method}
\title{Count the unique elements in data}
\usage{
countUnique(.data, ...)

\S4method{countUnique}{data.frame}(.data, ...)

\S4method{countUnique}{ANY}(.data, ...)
}
\arguments{
\item{.data}{A data.frame or similar.}

\item{...}{Columns in \code{dta}.}
}
\description{
Count the unique elements in data
}
