% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createIntervalsAge}
\alias{createIntervalsAge}
\title{Create age/positive number intervals from a variable.}
\usage{
createIntervalsAge(x, by, from = 0, to = 90, sep = "-")
}
\arguments{
\item{x}{A vector of numbers.}

\item{from}{Interval start.}

\item{to}{Interval end.}
}
\description{
Create age/positive number intervals from a variable.
}
\details{
Create age intervals
Each number in x will be placed in one of the intervals between arguments from
and to where by specifies the range of each interval. The right end of the interval
is by default set to Inf, i.e. oldest will be "anyone above 90" if by=90.
}
