% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbCreateTypeEnum}
\alias{dbCreateTypeEnum}
\title{Create a database ENUM type.}
\usage{
dbCreateTypeEnum(con, name, values)
}
\arguments{
\item{con}{A database connections.}

\item{name}{The name of the ENUM type.}

\item{values}{A vector of values in the ENUM.}
}
\description{
Create a database ENUM type.
}
