% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbRemoveTables}
\alias{dbRemoveTables}
\title{Remove database tables.}
\usage{
dbRemoveTables(con, tables)
}
\arguments{
\item{con}{A database connection.}

\item{tables}{A vector of tables to remove.}
}
\description{
Remove database tables.
}
