% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diffDays}
\alias{diffDays}
\title{Calculate difference between dates in days}
\usage{
diffDays(datesA, datesB)
}
\arguments{
\item{datesA}{A date/vector of dates.}

\item{datesB}{A date/vector of dates.}
}
\value{
\code{datesA}-\code{datesB} in days.
}
\description{
Calculate difference between dates in days
}
