% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandRange}
\alias{expandRange}
\title{Expand a numerical range embedded in a string}
\usage{
expandRange(x, sep = "-")
}
\arguments{
\item{x}{Strings to expand.}

\item{sep}{A separator ofr anges}
}
\description{
Expand a numerical range embedded in a string
}
\details{
\code{expandRange} will take strings in the format of "A#-A#" and repeat these
into range.
}
\examples{
vec <- c("A", "A1-A3")
vec <- expandRange(vec)
# vec is now c("A", "A1", "A2", "A3")

}
