% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frequencyCountDistinct}
\alias{frequencyCountDistinct}
\title{Count frequencies of discrete values.}
\usage{
frequencyCountDistinct(dta, col, ...)
}
\arguments{
\item{dta}{Any data accepted by dplyr.}

\item{col}{Column in dta to count values.}

\item{...}{Columns to select in data.}
}
\description{
Count frequencies of discrete values.
}
\details{
Count frequencies
Without any columns passed in (...) this function simply returns the distinct
values, counts and percent of these values. By passing other
columns through ... the function performs this operation over these columns so
that each value is only counted once for each group.
}
