% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{getColumn}
\alias{getColumn}
\title{Get column name using generic name}
\usage{
getColumn(.data, name)
}
\arguments{
\item{.data}{A dataRegister object.}

\item{name}{The generic name.}
}
\description{
Get column name using generic name
}
