% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{getExcelData}
\alias{getExcelData}
\title{Get data from an excelfile by searching through all sheets.}
\usage{
getExcelData(file, col, val, skip, colsSelect)
}
\arguments{
\item{file}{The excel file path.}

\item{col}{The column to search.}

\item{val}{The value to look for in col.}

\item{skip}{Number of lines to skip in each sheet.}

\item{colsSelect}{Which columns to select.}
}
\description{
Get data from an excelfile by searching through all sheets.
}
