% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glueCollapse}
\alias{glueCollapse}
\title{String interpolation with pre-collapsing}
\usage{
glueCollapse(..., sep = "")
}
\arguments{
\item{...}{Arguments to \code{glue::glue}.}

\item{sep}{Separator of values.}
}
\description{
String interpolation with pre-collapsing
}
