% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indexAlongUnique}
\alias{indexAlongUnique}
\title{Index unique observations}
\usage{
indexAlongUnique(dta, col, ...)
}
\arguments{
\item{dta}{A data frame or similar.}

\item{col}{The column to index}

\item{...}{Columns in \code{dta} defining}
}
\description{
Index unique observations
}
\seealso{
\code{\link[=indexAlong]{indexAlong()}}
}
