% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelStrings}
\alias{labelStrings}
\alias{labelStrings,character,character,character,ANY-method}
\alias{labelStrings,character,character,character,missing-method}
\alias{labelStrings,character,list,ANY,ANY-method}
\title{Match a vector against another character set and label matches.}
\usage{
labelStrings(strings, sets, labels, exclude)

\S4method{labelStrings}{character,character,character,ANY}(strings, sets, labels, exclude)

\S4method{labelStrings}{character,character,character,missing}(strings, sets, labels)

\S4method{labelStrings}{character,list,ANY,ANY}(strings, sets)
}
\arguments{
\item{strings}{A vector of strings.}

\item{sets}{A vector of strings to match against.}

\item{labels}{A vector of labels to assign to matches.}

\item{exclude}{A vector of strings to exclude. Either one for all sets or one per set.}
}
\description{
Match a vector against another character set and label matches.
}
\details{
\code{labelStrings} work similar to performing a left join, e.g., using \code{merge(x,y,all.x=T,...)}
if \code{y} is some data containing a label for a column existing in both \code{x} and \code{y}.
The difference is that this function allows for flexible matching (currently only
the starting characters in string), and allows for excluding patterns
For example, if all strings starting with "A" should have some label "Group A", but
not strings starting with "AB", "AB could be added as an exclusion pattern
(see example).
}
\examples{
x <- c("A", "ABC", "DEF")
labelStrings(x, "A", "Group A", "AB")

}
