% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{logging}
\alias{logging}
\alias{logCheckpoint}
\title{Logging}
\usage{
logExclusion(x, what, statistic, description)

logCheckpoint(x, what, statistic, description)
}
\arguments{
\item{x}{A \code{dataRegister} object.}

\item{what}{What is excluded?}

\item{statistic}{Number of exclusions.}

\item{description}{A description of the exclusion.}
}
\description{
Functions to log operations on data.
}
