% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{mergePeriods}
\alias{mergePeriods}
\title{Merge time periods}
\usage{
mergePeriods(dates, days, maxDistance = 0)
}
\arguments{
\item{dates}{Period starting date (YYYY-MM-DD)}

\item{days}{Period length in days}

\item{maxDistance}{Maximum distance between end and start of two periods for merging.}
}
\description{
Merge time periods
}
