% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R, R/utils-data.R
\name{mergePeriods}
\alias{mergePeriods}
\alias{getMergedPeriods}
\alias{getMergedPeriods,dataRegister-method}
\title{Merge dispensed drugs into non-overlapping treatment periods}
\usage{
getMergedPeriods(.data, ...)

\S4method{getMergedPeriods}{dataRegister}(.data, ...)

mergePeriods(dates, days, maxDistance = 0)
}
\arguments{
\item{.data}{Anything accepted by dplyr.}

\item{...}{arguments to \code{mergePeriods}.}

\item{dates}{Period starting date (YYYY-MM-DD)}

\item{days}{Period length in days}

\item{maxDistance}{Maximum distance between end and start of two periods for merging.}
}
\description{
Merge dispensed drugs into non-overlapping treatment periods

Merge time periods
}
\seealso{
\code{\link[=mergePeriods]{mergePeriods()}}
}
