% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R, R/utils-data.R
\name{getMergedPeriods}
\alias{getMergedPeriods}
\alias{getMergedPeriods,dataRegister-method}
\alias{mergePeriods}
\title{Merge dispensed drugs into non-overlapping treatment periods}
\usage{
getMergedPeriods(.data, .maxDistance = 0, .reset = F, .resetFun = mean)

\S4method{getMergedPeriods}{dataRegister}(.data, .maxDistance = 0, .reset = F, .resetFun = mean)

mergePeriods(dates, days, maxDistance = 0, reset = F, resetFun = mean)
}
\arguments{
\item{.data}{Anything accepted by dplyr.}

\item{.maxDistance}{Maximum distance between end and start of two periods for merging.}

\item{.reset}{Whether to add the overlapping time between two periods at then end of the merged period.}

\item{.resetFun}{A callable to use for dates with multiple days provided.}

\item{dates}{Period starting date (YYYY-MM-DD)}

\item{days}{Period length in days}

\item{maxDistance}{Maximum distance between end and start of two periods for merging.}

\item{reset}{Whether to add the overlapping time between two periods at then end of the merged period.}

\item{resetFun}{A callable to use for dates with multiple days provided.}
}
\description{
Merge dispensed drugs into non-overlapping treatment periods

Merge time periods
}
