% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{percent}
\alias{percent}
\title{Calculate percent based on counts}
\usage{
percent(.data, .colCount = "N", ...)
}
\arguments{
\item{.data}{Anything accepted by dplyr.}

\item{.colCount}{Name of column with counts.}

\item{...}{Grouping variables.}
}
\description{
Calculate percent based on counts
}
