% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{positionsAdjacent}
\alias{positionsAdjacent}
\title{Get positions of adjacent numbers}
\usage{
positionsAdjacent(numbers, .distance = 1)
}
\arguments{
\item{.distance}{Determinant of adjacent numbers.}

\item{.numbers}{A vector of numbers}
}
\description{
Get positions of adjacent numbers
}
