% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-registerdata.R
\name{renameColumns}
\alias{renameColumns}
\alias{renameColumns,dataRegister,data.frame-method}
\alias{renameColumns,data.frame,data.frame-method}
\title{Rename columns in a \code{dataRegister} object}
\usage{
renameColumns(.data, dtaCols, english = F, verbose = F)

\S4method{renameColumns}{dataRegister,data.frame}(.data, dtaCols, english = F, verbose = F)

\S4method{renameColumns}{data.frame,data.frame}(.data, dtaCols, english = F, verbose = F)
}
\arguments{
\item{.data}{A \code{dataRegister} object.}

\item{dtaCols}{A table with column names.}

\item{english}{Whether to use English names.}

\item{verbose}{Whether to output information.}
}
\description{
Rename columns in a \code{dataRegister} object
}
