% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-registers.R
\name{renameColumns}
\alias{renameColumns}
\alias{renameColumns,ANY,data.frame,logical-method}
\alias{renameColumns,ANY,data.frame,missing-method}
\title{Rename columns from delivery to new names}
\usage{
renameColumns(dta, dtaCols, english = F)

\S4method{renameColumns}{ANY,data.frame,logical}(dta, dtaCols, english = F)

\S4method{renameColumns}{ANY,data.frame,missing}(dta, dtaCols, english = F)
}
\arguments{
\item{dta}{A table with column names, e.g. data.frame.}

\item{dtaCols}{A data frame of column names}

\item{english}{Whether to translate column names to English}
}
\description{
Rename columns from delivery to new names
}
