% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{splitDataByRow}
\alias{splitDataByRow}
\title{Split data by rows that are equal to some value.}
\usage{
splitDataByRow(dta, splitIndicator = NA, header = F)
}
\arguments{
\item{dta}{Usually a data.frame.}

\item{splitIndicator}{A value to delineate rows that separate tables in data.}

\item{header}{If TRUE, the first row in each dataset are used as a header for the data.}
}
\description{
Split data by rows that are equal to some value.
}
\details{
The motivation for this function comes from spreadsheet in which multiple tables
are present in a sheet. If such tables are separated by some character or an entire
empty row this function can split such data into its constituents. If using
\code{header=TRUE} it is currently required that tables have the same columns.
}
\seealso{
\code{\link[=mergeDataByRow]{mergeDataByRow()}}
}
