% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txtNPercent}
\alias{txtNPercent}
\title{Create a string with n and percent}
\usage{
txtNPercent(n, N, strmask = "{n} ({percent})")
}
\arguments{
\item{n}{A number}

\item{N}{Another number}
}
\description{
Create a string with n and percent
}
